app_testtask42/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
Django==3.1.3
idna==2.10
isort==5.6.4
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.4
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
urllib3==1.25.11
vk-api==11.9.0
wrapt==1.12.1
